requirements.txt
fastapi
uvicorn
sqlmodel
httpx
python-dotenv
pytest
